Change float to decimal conversion, also allowing integers.

+++ users-service/users_item.py
@@ -6,4 +6,4 @@ from cassandra.cqlengine.models import Model
 
 class Users(Model):
     id = columns.UUID(primary_key=True, default=uuid.uuid4)
-    credit = columns.Float()
+    credit = columns.Decimal()
